Move scatterplot to own component as more maps statistics will follow
